Refacto of the code : split inference / send to annotation + orchestration (plug_to_pyroengine). Change of the sliding window to detect sequences (no overlapping). Change of how a sequence is considered containing enough images with wildfire for it to be sent to the annotation API.
